Restore picker on file:// pages on Chromium <106
Chromium MV3 ignores the "web_accessible_resources" manifest key on file://* schemes. This breaks the popup script, rendering the file picker unusable. This bug has been fixed in Chromium 106, which is not widely used yet.
https://crbug.com/1181450

(Warning: This will prevent the extension from working on file://* pages. Revert this change for use on Chromium 106+)